Move Profile to V1 namespace
