+++ spec/bambora/client_spec.rb
@@ -4,8 +4,21 @@ require 'spec_helper'
 
 module Bambora
   describe Client do
-    describe '#profiles' do
-      pending
+    describe '#profile' do
+      context 'with defaults' do
+        subject { Bambora::Client.new }
+
+        it 'returns a Bambora::V1::Profile' do
+          expect(subject.profile).to be_a Bambora::V1::Profile
+        end
+      end
+
+      context 'with a different version' do
+        subject { Bambora::Client.new(version: 'V2') }
+
+        it 'returns a raises an error' do
+        end
+      end
     end
 
     describe '#payments' do
